requirements.txt
jinja2>=3.0.0
PyYAML>=6.0
tabulate>=0.9.0
